Flytt form del av kode til egen fil før kode tilpasses
